requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
passlib[bcrypt]
pyjwt
pydantic
pydantic-settings
alembic
httpx
pytest
pytest-asyncio
python-multipart
